qtribe-platform/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-redis>=5.4,<6.0
django-filter>=23.5,<24.0
Pillow>=10.0,<11.0
PyMySQL>=1.1,<2.0
python-dotenv>=1.0,<2.0
redis>=5.0,<6.0
drf-spectacular>=0.27,<0.28
gunicorn>=21.0,<22.0
whitenoise>=6.6,<7.0
